added navigation to user management moved home and userprofile to dir
