--- app/src/main/java/com/mad/mad_encounters/Home.kt
@@ -1,11 +1,0 @@
-package com.mad.mad_encounters
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class Home : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_home)
-    }
-}
