+++ app/src/main/java/com/mad/mad_encounters/usermanagement/activity/UserProfile.kt
@@ -1,7 +1,8 @@
-package com.mad.mad_encounters
+package com.mad.mad_encounters.usermanagement.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.mad.mad_encounters.R
 
 class UserProfile : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
