+++ app/src/main/res/layout/activity_delete_order.xml
@@ -95,7 +95,8 @@
         android:layout_marginTop="10dp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="@+id/textView14"
-        app:srcCompat="@drawable/baseline_arrow_back_24" />
+        app:srcCompat="@drawable/baseline_arrow_back_24"
+        android:onClick="navigateback" />
 
     <ImageView
         android:id="@+id/imageView8"
