+++ app/src/main/res/layout/activity_user_profile.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserProfile">
+    tools:context=".usermanagement.activity.UserProfile">
 
     <TextView
         android:id="@+id/textView6"
