+++ app/src/main/res/layout/activity_view_order.xml
@@ -37,7 +37,8 @@
         android:layout_marginTop="10dp"
         app:layout_constraintStart_toStartOf="@+id/textView7"
         app:layout_constraintTop_toTopOf="@+id/textView7"
-        app:srcCompat="@drawable/baseline_arrow_back_24" />
+        app:srcCompat="@drawable/baseline_arrow_back_24"
+        android:onClick="navigateback" />
 
     <ImageView
         android:id="@+id/imageView5"
